+++ R/15_Biclustering_nominal.R
@@ -194,6 +194,7 @@ Biclustering.nominal <- function(U,
   # output ----------------------------------------------------------
   cls <- apply(clsmemb, 1, which.max)
   fld <- apply(fldmemb, 1, which.max)
+  check_empty_fields(fld, nfld)
   fldmemb01 <- sign(fldmemb - apply(fldmemb, 1, max)) + 1
   flddist <- colSums(fldmemb01)
   clsmemb01 <- sign(clsmemb - apply(clsmemb, 1, max)) + 1